Refer to HTML 5.1 _REC_, add HTML 5.2 ED
The URI of the former HTML 5.1 draft now refers to the HTML 5.2+ working draft which breaks section links (e.g. `{{SpecName('HTML5.1', '#dom-windowbase64-atob', 'WindowBase64.atob()')}}`).

Refer to the HTML 5.1 Recommendation instead.
(Approach recommended in comment on top is questionable: MDN links to ED, but said “REC” in next column.)

References to sections in SpecName() calls need to be updated.