Updated the cssxref() macro to simplify the code that builds the link text and to fix a bug causing the text to be forced to lower-case even though some CSS terms are mixed- case (like translateX()).
Also revised the logic of building the summary and link text to still fetch the page
summary for the tooltip when the link text parameter is included, and to only wrap
the link text in <code>..</code> if the link text parameter wasn't included, and
if the link text has no spaces; this lets cssxref be used in some cases where the
link text isn't an API term but is just a plaintext link to the reference page.

The updated code now just grabs the page.title of the destination if the page exists,
otherwise it just uses the $0 or $1 that was originally provided.

The title and summary are both escaped before use, with the latter having HTML tags
removed as well.