Corrected the typos
I did not know about the fix. I asked on the IRC channel, since I want to translate some of the texts into German and was directed to this file. I have seen that the menus are not the same on the English and the German version. It's very confusing at the moment when you switch between the two languages.