Updated HTTPSidebar.ejs to enhance organization
This patch adds the "Protocol upgrade mechanism" document to the sidebar in
the HTTP section of MDN. It also updates the organization a bit; the article
"HTTP Basics" was previously unavailable as its link was only on a list item
used as a group title. Now that group title is changed to "HTTP guide" and
"HTTP Basics" is the first entry beneath. This also serves to let the
"Protocol upgrade mechanism" article lie underneath without seeming too
complex for a section labeled "basics."