Beef up article menu macros
Updates the macros that generate definition list-based article lists to
escape things correctly do avoid potentially treating text describing
HTML elements from being treated as actual HTML while rendering the
page. I had seen this happen in a few tests.

In addition, the structure of several of these macros is much improved,
and the construction of the two-column layouts should be more reliable
and attractive.

There's also now a new SubpagesWithTagsOneCol macro that generates a
list of subpages matching any of the specified tags, but in a single
column. The original SubpagesWithTags macro still does a two-column
output, by wrapping a call to the OneCol version in a call to the
MakeColumnsForDL macro that converts a definition list to two-column
format.