Refactor SubpagesWithSummaries and MakeColumnsForDL
This patch refactors the SubpagesWithSummaries and MakeColumnsForDL
macros. This is done together since they are somewhat interdependent
given that the output from SubpagesWithSummaries is fed into the other.

The changes to MakeColumnsForDL strengthen the code that splits a <dl>
into two columns to avoid cases where the second column is
unnecessarily longer than the first. This is done by computing the
column midpoint based on the text content of the HTML rather than
the HTML itself, by stripping tags from the content.

The column splitting is also capable of handling lists where the <dl>,
<dd>, and <dt> tags have unexpected attributes.

The changes to SubpagesWithSummaries clean up the logic in general
and handle proper escaping of the output to avoid cases where text
describing HTML tags can result in those tags being inserted as
raw HTML into the code, breaking content.

In addition, I've clarified the code to be easier to read and update in
the future, and have added numerous comments to help explain what
is going on.