Refactor SubapgesWithTags, add SubpagesWithTagsOneCol
This patch moves the logic from SubpagesWithTags into SubpagesWithTagsOneCol,
which outputs a one-column list of subpages which feature one or more of
a set of tags. Then SubpagesWithTags is simply piping the output from that
into MakeColumnsForDL to create a two-column output.

This also refactors SubpagesWithTagsOneCol to use the revised and improved
logic from the SubpagesWithSummaries refactor done recently. In the future
we can improve this further to let these all share the same code
but trying to do this a atep at a time.