Update HTMLSidebar macro for new HTML element page titles
For SEO reasons, the element pages' titles are being changed to the
form:

`<elementname>`: The Element Name element

This patch updates `HTMLSidebar` to strip off everything that isn't
the element name, so the title is what's expected by other parts of
the site. This lets the sidebar's list of HTML elements be correctly
formatted.

THIS PATCH REQUIRES THE `ListSubpagesForSidebar` PATCH (PR #401).