Update ListSubpagesForSidebar macro to allow trimming with delimiters
This patch adds two new parameters to `ListSubpagesForSidebar`. These
let you specify start and end delimiters, so that the titles of pages
are cropped to include only text that's located between those
characters.

This lets us, for instance, build a list with the items' titles
including only the name of the HTML element in the element reference,
even when the page titles include other text.

For example:

`{{ListSubpagesForSidebar("/en-US/docs/Web/HTML/Element", 0, 0, "<", ">")}}`

This builds a sidebar list of the HTML element reference pages, with only
the elements' names included in each link, instead of the entire title of
the pages.