Update instructions to regen npm-shrinkwrap.json
The Docker build fails if npm-shrinkwrap.json is deleted, because it
attempts to copy a file that doesn't exist. Instead, replace the
contents with an empty dictionary, so that "npm install" uses fresh
versions of dependencies.

Also, use KS_VERSION=latest and VERSION=latest, so that the
docker-compose files don't have to be modified to use the locally built
image.