Fix to better pt-BR translation
There's no correct way to translate **layout** to Brazilian Portuguese. The thing is layout may refer to the design of an interface or also to the schema, in a more technical way. In this PR I use the schema as a translation, because the content of the theme is closest to the technical approach, instead of talking about usability or design terms.