bug 1482159: Add sample server URL to config
Add the live sample base URL as an environment-overridable configuraton,
and use it in the LiveSampleURL macro to switch to the live sample
server hostname. The default configuration is the production value.

The tests for EmbedLiveSample also need this config.

This also switches to the WHATWG URL class, which properly encodes the
URLs.