Add trailing slashes to several spec URLs
This change adds a trailing slash to the `url` values for the Media
Capabilities, XMLHttpRequest, and Compatibility specs in the
SpecData.json file — making the formatting of the `url` values for those
specs consistent with the formatting of other specs’ `url` values, which
all have trailing slashes.

Without this change, MDN articles end up with URLs in their
Specifications sections that look like this:

https://compat.spec.whatwg.org#the-webkit-text-stroke-color

...which when navigated to, redirects to this:

https://compat.spec.whatwg.org/#the-webkit-text-stroke-color

So, adding the trailing slash to the `url` value in the SpecData.json
file eliminates the need for those redirects to happen.