Bug 1505084: update KumaScript to use Node 10.14.2
This commit just updates the package.json file in the KumaScript repo
and regenerates the corresponding npm-shrinkwrap.js file. Along with the
Node verison update, we update the version of the fibers package because
the old version we were using has a native component that would no longer
compile with Node 10.

The Kuma repo contains the Dockerfile that creates the Docker image
for KumaScript. When Kuma is updated to use this revision of KumaScript
we must update the Dockerfile at the same time or we will have a
version mismatch between the version of node specified in package.json
and the version actually installed in the KumaScript container.

Test Plan:
- tested following the directions at
  https://github.com/davidflanagan/kumascript#updating-the-dockerfile-andor-packagejson-file
- ran kuma locally without problems