Bug 1522206: Define translation text for the WebExtAPISidebar macro.
The WebExtAPISidebar macro (used by AddonSidebar) tries to access
a text[] array that does not exist. This appears to be an old bug
that only recently started throwing an error after the KumaScript
refactor.

This PR takes the text array from the APIRef macro and uses a
tweaked version in WebExtAPISidebar. It also adds translations
to the sidebar strings "Properties", "Methods", "Types" and "Events".