fix(macros): Correct {{DOMxRef(…)}} URL computation (#1074)
The domxref macro converts its first argument to a URL component, and as part of this
converts strings to underscores. But it was only converting the first string to an underscore.
This patch fixes that so that all spaces are converted to underscore. It also does some minor
formatting cleanups to remove trailing spaces at the ends of lines.