Refactor the various macro APIs in environment.js into separate api/*.js files.
This patch takes lots of macro API code out of src/environment.js and puts it into individual
API-specific files in src/api/*.js, replacing the code in environment.js with a simple require()
call. This just moves the code: none of the code actually changes.

This improves modularity and also separates the modern, well-tested code in environment.js from the legacy, untested code in api/*.js. In particular, it brings the test coverage for environment.js up
to 100%.