GroupData: WebXR Device API to latest spec (#1253)
* GroupData: WebXR Device API to latest spec
This PR makes the changes needed to bring the
WebXR Device API spec up to the current state of
the specification. Also moves some items which
were incorrectly listed as "types" into the
"dictionaries" section for WebXR.

* Add XREnvironmentBlendMode to GroupData
Added the XREnvironmentBlendMode type to GroupData.
It's part of the WebXR Augmented Reality standard
and I've decided to merge all WebXR standards together
in GroupData, since they're so closely connected.